Update YGB-master with master

+++ docroot/modules/contrib/queue_ui/src/QueueUIBatch.php
@@ -25,6 +25,7 @@ class QueueUIBatch {
     $queueFactory = \Drupal::service('queue');
 
     $info = $queueManager->getDefinition($queue_name);
+    $title = $info['title'];
 
     // Make sure every queue exists. There is no harm in trying to recreate
     // an existing queue.
@@ -32,28 +33,32 @@ class QueueUIBatch {
 
     $queue_worker = $queueManager->createInstance($queue_name);
     $queue = $queueFactory->get($queue_name);
-    $item = $queue->claimItem();
 
-    try {
-      $queue_worker->processItem($item->data);
-      $queue->deleteItem($item);
+    if ($item = $queue->claimItem()) {
+      try {
+        $queue_worker->processItem($item->data);
+
+        // Let other modules alter the title of the item being processed
+        \Drupal::moduleHandler()->alter('queue_ui_batch_title', $title, $item->data);
+        $context['message'] = $title;
+
+        $queue->deleteItem($item);
+      }
+      catch (SuspendQueueException $e) {
+        // If the worker indicates there is a problem with the whole queue,
+        // release the item and skip to the next queue.
+        $queue->releaseItem($item);
+
+        watchdog_exception('cron', $e);
+
+        // Skip to the next queue.
+        $context['interrupted'] = TRUE;
+      }
+      catch (\Exception $e) {
+        // In case of any other kind of exception, log it and leave the item
+        // in the queue to be processed again later.
+        watchdog_exception('cron', $e);
+      }
     }
-    catch (SuspendQueueException $e) {
-      // If the worker indicates there is a problem with the whole queue,
-      // release the item and skip to the next queue.
-      $queue->releaseItem($item);
-
-      watchdog_exception('cron', $e);
-
-      // Skip to the next queue.
-      $context['interrupted'] = TRUE;
-    }
-    catch (\Exception $e) {
-      // In case of any other kind of exception, log it and leave the item
-      // in the queue to be processed again later.
-      watchdog_exception('cron', $e);
-    }
-
-    $context['message'] = $info['title'];
   }
 }
